chore(api-tests): move sourcemap_helpers + RUM ingest scripts out of tests/
- sourcemap_helpers.py: utility used only by test_sourcemap_api.py.
  Moved to helpers/ (the conventional location for shared, non-test code).
- ingest_rum_errors.py + ingest_rum_performance.py: standalone scripts
  (no `def test_*`, never imported by any test). Moved to fixtures/rum/
  alongside other test data.

Verified: pytest collection still discovers 517 tests; test_sourcemap_api
imports resolve correctly (still skipped on OSS as before).
